autogen/agentchat/contrib/captainagent/tools/requirements.txt
markdownify
arxiv
pymupdf
wikipedia-api
easyocr
python-pptx
openai-whisper
pandas
scipy
sentence-transformers
